Change bash and zsh prompts to delay eval
In the bash/zsh init scripts, PS1/PROMPT are being evaluated
once (when the variable is assigned to) because double quotes
do not sufficiently defer command substitution. This leads to
the prompt getting "stuck" wherever starship was initialized.

This changes the init script to use single quotes, deferring
command substitution until the prompt variable is evaluated.