Reduce scope of test char_module_vicmd_keymap
It was unnecessary (in fact, problematic) for this test to be concerned
with matching the entire output of the module; in this case, because
different behaviour is expected for different values of STARSHIP_SHELL,
and there is no way to mock the shell ANSI escape logic out of test.

The solution is a slightly less *exacting* test that does a simple
substring match for the expected character against the output - at
least, until such a time as mocking the ANSI escaping out becomes
viable.