fix: Escape data that bash can execute
When PS1 includes something like '$(foo)' it will be evaluated. This
means that if the value of `starship prompt` includes such a construct,
that will be evaluated. It is possible for malicious parties to use this
behavior to execute code. See https://github.com/njhartwell/pw3nage for
a proof of concept.

This change sets the PS1 variable to call starship itself, so the
'$(foo)' construct returned by starship will not be evaluated by bash.