feat: Add support for the ion shell
This commit adds support for the jobs and status options of the starship prompt.
Support for the cmd-duration is also added if you are using the latest
ion shell. The cmd-duration is implemented using the ION_CMD_DURATION variable.
This variable keeps track of the time spent executing the last command. (This is
similar to the CMD_DURATION of the fish shell).

See this MR for more information: https://gitlab.redox-os.org/redox-os/ion/merge_requests/1176