proof of concept for showing the exit status
I'd like a way to automatically print the exit status of failed
commands, in addition to the fact that they failed. Here's an initial
stab at enabling something like this. (I haven't updated tests yet, so
for now this change causes failures.) With this change, I add this to my
config:

    [character]
    error_format = "[$code](red) [❯](bold red) "

Then my prompt looks like this (but with color):

    starship on  master
    ❯ true

    starship on  master
    ❯ false

    starship on  master
    1 ❯ bash -c "exit 4"

    starship on  master
    4 ❯

I'm not sure if this is the best approach, either design-wise or
implementation-wise, but I wanted to put something up to get
suggestions. Another option would be to print an extra line above the
prompt containing the status; that might make is clearer that it's
associated with the previous command too.