fix: Change nodejs symbol to the nerd-font nodejs one
The unicode hexagon symbol does not actually fit into a single column with a fixed-width font. [ (nf-mdi-nodejs)](https://www.nerdfonts.com/cheat-sheet?set=nf-mdi-nodejs) does. As starship requires a nerd font as a prerequisite, it's safe to assume that this symbol is available.

In the documentation, I've left the '⬢' symbols in '#### Text Group' as
they are, as they will continue displaying correctly in a browser
without a nerd font available. I feel like readability is more important
than consistency with the new nodejs symbol, especially as this
documentation section does not actually refer to nodejs, rather it's
just a symbol.