ci(deploy): Fix failing Linux GNU builds
This fixes the failing Linux 64 bit builds that are linking to GNU libc
that are failing because cross doesn't compile openssl-sys. The fix is
simply to enable the vendored-tls feature so that we get the rust
vendored SSL rather than the system SSL. Have also done a little bit of
refactoring to the deploy script. Now rather than have a matrix of
targets have just included each target individually (this may be
something we want to revisit in the future as the number of platforms
increases). I have also removed the installation of the musl tools since
this is no longer needed when using cross for compilation.