fix(directory): Remove re-implementation of "current_dir" resolution logic
The `directory` module function `get_current_dir` was a subtly different implementation of the same routine in `Context::new`.
The difference in behavior leads it to ignore the `--path` override argument (sort of - the "is_readonly_dir" test was run against the `context.current_dir` value already).

The difference is in the priority order of the three sources checked for the "current directory":

- `Context::new`
    - Try `--path` argument
    - Or try `PWD` environment variable
    - Or try `env::current_dir` routine

- `get_current_dir`
    - Check `use_logical_path` config
        - If true, try `PWD` environment variable
        - If false, try `env::current_dir` routine
    - Or try `--path` argument

This commit removes the `get_current_dir` function from the directory module and instead we now rely on the `current_dir` value already written to the Context.
The "powershell" prefix cleanup code has additionally been ported into `Context::new_with_dir`, and the unit tests updated.