refactor: install.sh to improve usage and install instructions (#3025)
* Refactored the usage function to use printf

Using printf to print the usage function instead of a here doc handles
the formatting for us so it will be consistent across a wide variety of
terminal emulators. Here docs also tend to mess up auto-formatters so
removing it has that added benefit.

* Made the install instructions their own function

The function loops through simular install instructions so this way we
won't have to repeat ourselves every time we add a new supported shell.

* Set default config_file location

The default location is based on the name of the shell. It is overwritten when needed.

Co-authored-by: Thomas O'Donnell <andytom@users.noreply.github.com>

Co-authored-by: Thomas O'Donnell <andytom@users.noreply.github.com>