feat: add npm module
This adds the npm module so that when in a directory with a
package-lock.json or an .npmrc, you will see the npm version displayed.
Usually this will be after the nodejs module because by default the
nodejs module will display if there is a package.json and where there is
a package-lock.json, there is usually a package.json.

The npm module is disabled by default because it can add extra unwanted
noise to your prompt. However it can also be very useful to have in your
prompt because npm is known for having different behaviors between major
versions that will subtly break your project.

This also mimicks the nodejs module when it comes to checking the
engines field in package.json. The default color is a dark red, and if
the current npm version is incompatible with the engines field in the
local package.json, it will display as the default red, to match the
color of the nodejs module when it is incompatible. Yes, npm's color is
red and when there is an error it is just a lighter red, but I didn't
know what color to make npm by default and still have an error color
that made sense.