feat: Add support for Daml
#### Description
- Add a `daml` module that reads the currently used SDK version for a project using the one indicated in the `daml.yaml` file or overrides it with `DAML_SDK_VERSION`
- Add Daml support to the `package` module (uses the `version` key in the `daml.yaml` file)

#### Motivation and Context
Closes #3903

#### How Has This Been Tested?
- [x] I have tested using **MacOS**
- [ ] I have tested using **Linux**
- [ ] I have tested using **Windows**

Apart from the relevant automated tests, I opened a sample Daml project and touched the `sdk-version` and `version` keys in the `daml.yaml` file, as well as the `DAML_SDK_VERSION` environment variable to see whether the changes were picked up correctly. Also went into a sub-directory to see whether the information disappeared as expected. I further made sure that the example in the documentation works, as well as checking that every option works as expected.

#### Checklist:
- [x] I have updated the documentation accordingly.
- [x] I have updated the tests accordingly.