docs(config): Replace misleading $all reference
In the config `$all` has special meaning, as it is and contains the default value for `format`.

Using `$all` in the example for conditional format strings with multiple variables is thus potentially misleading and confusing.
Using a neutral variable `$combined` has no conflict with other meanings and is thus preferable.