feat(preset): Add No Empty Icons preset (#4518)
* feat(preset): Add No Empty Icons preset

When toolset files are identified by the Starship module, the default format displays the toolset icon, and additional information.

When the toolset executables are not available, the additional information (like version number) is missing. Only the toolset icon is displayed.

The No Empty Icons preset changes the format configuration to not show the toolset icons if the variables are empty - presumably because the toolset is not installed or found.

Closes #3070
Related #3248

* Remove non-version-related modules from preset

Co-authored-by: David Knaack <davidkna@users.noreply.github.com>

* Add new module opa to preset

Co-authored-by: David Knaack <davidkna@users.noreply.github.com>

* Drop unnecessary inner conditional from format

Co-authored-by: David Knaack <davidkna@users.noreply.github.com>

* Remove commit_hash_length setting from preset

* Remove undesired modules

Co-authored-by: David Knaack <davidkna@users.noreply.github.com>

Co-authored-by: David Knaack <davidkna@users.noreply.github.com>