Make STARSHIP_PREEXEC_READY work in 'strict mode'
I am sourcing Starship through a script that uses 'strict mode'  (`set -o nounset`) to exit if an var is unbound.  `STARSHIP_PREEXEC_READY` is unbound if not yet set by `starship_precmd`. 

This PR changes the behavior so that `STARSHIP_PREEXEC_READY` falls back to an empty bound var if the var is not yet bound.