feat: Add `opt_in_modules` config option
When `true`, it turns module selection into opt-in (all modules are
disabled by default, with some exceptions), instead of opt-out as it is
now.

Technically, if `opt_in_modules` is true, the default value for any
`disabled` module key will then be false.

The default value for this new option is `false` meaning the default
value for any `disabled` module key will be true, as it is now.