feat(line_break): add break_below_width option
When specified, this option will change `line_break`'s behaviour to only
insert a line break when the width of the terminal is below the given
value.  This allows users to increase the vertical space used by their
prompt only when horizontal space is at a premium.

(This is particularly useful when a right prompt is configured, as a
sufficiently long left prompt may mean it is never displayed in smaller
terminals with `line_break` completely disabled.)