fix: assure the most important settings are picked up by `gitoxide`.
Now that `gitoxide` is used for diffing, there are a lot of important
settings. `core.autocrlf` is among them, which on Windows greatly affects
the outcome of the `git_metrics` module. This also means we have to run
the Git binary to get its system configuration location, which is costly
particularly on Windows where it is most direly needed.