docs(preset/*-powerline): Remove "rainbow effect"
With pastel-powerline/catppuccin-powerline, when there are modules with
no output, the  characters is always printed, resulting in something
like rainbow when there are several empty modules.

This commit moves this  character to the individual module `format`
directive, and uses `prev_bg` as its background to get the correct
color.

This also makes it easier to change the color of each prompt section as
the colors choices are all done in the module configuration.

Signed-off-by: Christophe Fergeau <christophe@fergeau.eu>