Filter out non-printing ANSI escape codes
The string `\x1b[35;49mkai` is, on the screen, only 3 characters long,
however the `width_graphemes()` function would say that it is 10
characters long. This is only really an issue for `explain`, as it makes
the width of the left column incorrect for the padding. Instead of the
following:

```console
$ starship explain
"/" (<1ms)    - <description>
"kai" (<1ms)  - <description>
```

it would print out like this for any parts that don't use ansi escapes
internally:

```console
"/" (<1ms)           - <description>
"kai" (<1ms)  - <description>
```

This should cause no problems with any of the unstyled lengths that
are generated normally, but any pre-styled strings in `custom` calls
will cause this issue.